ai_service/requirements.txt
fastapi==0.115.0
uvicorn==0.31.1
supabase==2.10.0
numpy==2.1.2
pandas==2.2.3
scikit-learn==1.5.2
python-dotenv==1.0.1
pydantic==2.9.2
joblib==1.4.2
scipy==1.14.1
plotly==5.24.1
prophet==1.1.6
